backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
click==8.1.8
coverage==7.8.0
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
django-rq==3.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
nose==1.3.7
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
redis==5.2.1
rq==2.1.0
sentry-sdk==2.27.0
sqlparse==0.5.3
urllib3==2.4.0
whitenoise==6.9.0
